Move tests; Add tests to all_tests file
