+++ ssv/spectest/all_tests.go
@@ -73,6 +73,8 @@ var AllTests = []tests.TestF{
 
 	proposer.ProposeBlindedBlockDecidedRegular,
 	proposer.ProposeRegularBlockDecidedBlinded,
+	proposer.BlindedRunnerAcceptsNormalBlock,
+	proposer.NormalProposerAcceptsBlindedBlock,
 
 	// pre_consensus_justifications.PastSlot,
 	// pre_consensus_justifications.InvalidData,
